// api/types/pagination.ts
export interface IPaginatedResponse<T> {
    current_page: number;
    data: T[];
    first_page_url: string;
    from: number;
    last_page: number;
    last_page_url: string;
    next_page_url: string | null;
    path: string;
    per_page: number | string;
    prev_page_url: string | null;
    to: number;
    total: number;
}